Cleanup preview param
