--- buildSrc/build.gradle
@@ -1,9 +1,0 @@
-repositories {
-  mavenCentral()
-  google()
-}
-
-dependencies {
-  implementation libs.plugin.android
-  implementation libs.plugin.kotlin
-}
