--- buildSrc/settings.gradle
@@ -1,9 +1,0 @@
-rootProject.name = 'build-support'
-
-dependencyResolutionManagement {
-  versionCatalogs {
-    libs {
-      from(files('../gradle/libs.versions.toml'))
-    }
-  }
-}
