+++ paparazzi-annotations/build.gradle
@@ -1,10 +1,7 @@
 apply plugin: 'org.jetbrains.kotlin.jvm'
 apply plugin: 'org.jetbrains.kotlin.plugin.compose'
 apply plugin: 'com.vanniktech.maven.publish'
-apply plugin: 'aar2jar'
-
-aar2jar.configureForConfiguration("compileOnly")
 
 dependencies {
-  compileOnlyAarAsJar libs.compose.runtime
+  compileOnly libs.compose.runtime
 }
