+++ paparazzi-preview-test-junit/src/main/java/app/cash/paparazzi/preview/Snapshot.kt
@@ -3,14 +3,19 @@ package app.cash.paparazzi.preview
 
 import app.cash.paparazzi.Paparazzi
 import app.cash.paparazzi.preview.runtime.PaparazziPreviewData
-import com.google.testing.junit.testparameterinjector.TestParameterValuesProvider
+import com.google.testing.junit.testparameterinjector.TestParameter.TestParameterValuesProvider
 
 /**
  * Take a snapshot of the given [previewData].
  */
-public fun Paparazzi.snapshot(previewData: PaparazziPreviewData, name: String? = null) {
+public fun Paparazzi.snapshot(
+  previewData: PaparazziPreviewData,
+  name: String? = null
+) {
   when (previewData) {
     is PaparazziPreviewData.Default -> snapshotDefault(previewData, name)
+    is PaparazziPreviewData.Empty -> Unit
+    is PaparazziPreviewData.Error -> error(previewData.message)
   }
 }
