+++ paparazzi-preview-test-junit/src/main/java/app/cash/paparazzi/preview/Utils.kt
@@ -4,7 +4,10 @@ package app.cash.paparazzi.preview
 import app.cash.paparazzi.Paparazzi
 import app.cash.paparazzi.preview.runtime.PaparazziPreviewData
 
-internal fun Paparazzi.snapshotDefault(previewData: PaparazziPreviewData.Default, name: String?) {
+internal fun Paparazzi.snapshotDefault(
+  previewData: PaparazziPreviewData.Default,
+  name: String?
+) {
   snapshot(name) {
     previewData.composable()
   }
