+++ paparazzi/build.gradle
@@ -48,12 +48,10 @@ dependencies {
   api libs.guava
   api libs.kotlinx.coroutines.android
   api libs.okio
-  api libs.testParameterInjector
   api platform(libs.kotlin.bom)
   implementation libs.moshi.core
   implementation libs.moshi.adapters
   implementation libs.moshi.kotlinReflect
-  implementation projects.paparazziAnnotations
 
   def osName = System.getProperty("os.name").toLowerCase(Locale.US)
   def osLabel
